requirements.txt
# Основные библиотеки
aiohttp==3.9.1
python-dotenv==1.0.0
SQLAlchemy==2.0.25
pydantic==2.5.3
telethon==1.32.1

# Библиотеки для MySQL
pymysql==1.1.0
cryptography==42.0.0

# Библиотеки для обработки данных
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.3.2

# Библиотеки для работы с OpenAI
openai==1.12.0

# Библиотеки для логирования и управления процессами
watchdog==3.0.0
supervisor==4.2.5

# Библиотеки для тестирования
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.11.1
